# openapi.json
{"openapi":"3.1.0","info":{"title":"HAIOS API","version":"0.1.0"},"paths":{"/api/v1/sessions":{"post":{"summary":"Create Session","description":"Create a new session and return a signed JWT.","operationId":"create_session_api_v1_sessions_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SessionCreateRequest"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SessionCreateResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/chat":{"get":{"summary":"Chat Endpoint Health","description":"Chat service health check.","operationId":"chat_endpoint_health_chat_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}},"post":{"summary":"Chat Endpoint","description":"Chat endpoint for frontend interaction.","operationId":"chat_endpoint_chat_post","requestBody":{"content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Request"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/health":{"get":{"summary":"Health Check","description":"System health check endpoint.","operationId":"health_check_api_v1_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HealthResponse"}}}}}}},"/api/v1/providers":{"get":{"summary":"List Providers","description":"List all registered providers and their current status.","operationId":"list_providers_api_v1_providers_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Response List Providers Api V1 Providers Get"}}}}}}},"/metrics":{"get":{"summary":"Prometheus Metrics","description":"Expose Prometheus metrics.","operationId":"prometheus_metrics_metrics_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/":{"get":{"summary":"Root","description":"Root endpoint.","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Response Root  Get"}}}}}}}},"components":{"schemas":{"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"HealthResponse":{"properties":{"status":{"type":"string","title":"Status","default":"ok"},"version":{"type":"string","title":"Version"},"uptime_seconds":{"type":"number","title":"Uptime Seconds"}},"type":"object","required":["version","uptime_seconds"],"title":"HealthResponse","description":"Health check response."},"SessionCreateRequest":{"properties":{"user_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"User Id"},"metadata":{"additionalProperties":true,"type":"object","title":"Metadata"}},"type":"object","title":"SessionCreateRequest","description":"Request to create a new session."},"SessionCreateResponse":{"properties":{"session_id":{"type":"string","title":"Session Id"},"token":{"type":"string","title":"Token"},"expires_in":{"type":"integer","title":"Expires In","description":"Token TTL in seconds"}},"type":"object","required":["session_id","token","expires_in"],"title":"SessionCreateResponse","description":"Response containing the new session token."},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}